Create topic controller
